requirements.txt
fastapi 
uvicorn 
gunicorn 
tensorflow-cpu
nltk==3.8.1
unidecode 